adds new routes spec and some new tests, also starts change to BigDecimal

+++ src/main/scala/com/example/Model.scala
@@ -48,7 +48,7 @@ object ReviewSummary {
     deriveDecoder[ReviewSummary]
 }
 
-case class Result(asin: String, averageRating: Double)
+case class Result(asin: String, averageRating: BigDecimal)
 
 object Result {
   implicit val resultEncoder: Encoder[Result] = deriveEncoder[Result]
